==> 1. Get categories for each champion.ipynb <==
import os
import re
import scrapy
import shutil
import scrapy
from bs4 import BeautifulSoup
from scrapy.crawler import CrawlerProcess
import requests
from urllib.parse import unquote
from time import sleep
import pandas as pd


choices = ['https://leagueoflegends.fandom.com/wiki/Category:Artillery_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Assassin_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Battlemage_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Burst_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Catcher_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Controller_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Diver_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Enchanter_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Fighter_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Juggernaut_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Mage_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Marksman_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Skirmisher_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Slayer_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Specialist_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Tank_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Vanguard_champion',
 'https://leagueoflegends.fandom.com/wiki/Category:Warden_champion']
choice1 = choices[0]
choice1


final_answers = {}
for choice in choices:
    category = choice[:-9].split(":")[-1]
    r = requests.get(choice)
    answers = [unquote(x)[5:][:-4] for x in set(re.findall("wiki/[^/]+/LoL", r.text))]
    final_answers[category] = answers
    sleep(3)
final_answers


rows

my_lookup_ids = {"Artillery": 6,
"Assassin": 11,
"Battlemage": 10,
"Burst": 6,
"Catcher": 8,
"Diver": 2,
"Enchanter": 4,
"Juggernaut": 3,
"Marksman": 5,
"Skirmisher": 9,
"Specialist": 0,
"Vanguard": 1,
"Warden": 7}

redefine_roles_specialists_or_ambiguous = {"Gnar": 9,
"Azir": 6,
"Fiddlesticks": 2,
"Teemo": 10,
"Heimerdinger": 6,
"Nidalee": 6,
"Quinn": 5,
"Singed": 1,
"Graves": 5,
"Cho'Gath": 7,
"Kayle": 5,
"Zilean": 7,
"Kennen": 6,
"Gangplank": 9,
                              
                              
                              
"Diana": 11,
"Jhin": 5,
"Karma": 4,
"Lux": 6,
"Neeko": 6,
"Pyke": 11,
"Rengar": 11,
"Senna": 4,
"Seraphine": 4,
"Sylas": 6,
"Taric": 7,
"Varus": 5,
"Yone": 9,                             
                             
                             }




classes = []
champions = []
for (x, y) in final_answers.items():
    x = my_lookup_ids.get(x, -1)
    for champion in y:
        if champion in redefine_roles_specialists:
            x = redefine_roles_specialists_or_ambiguous[champion]
        classes.append(x)
        champions.append(champion)
df = pd.DataFrame()
df['class'] = classes
df['champion'] = champions
df = df.drop_duplicates(subset=['champion'])
df


df.to_csv("categories.csv", index=False)
