==> 2. Find Out What Pros Are Playing.ipynb <==
import os
import re
import scrapy
import shutil
import scrapy
from bs4 import BeautifulSoup
from scrapy.crawler import CrawlerProcess
import requests
from urllib.parse import unquote
from time import sleep
import pandas as pd
from glob import glob


url_id = 33878
urls = []
for i in range(100):
    url_id -= 1
    url = f"https://gol.gg/game/stats/{url_id}/page-game/"
    # print(url)
    urls.append(url)


# r = requests.get(url)
# r.text


import webbrowser
import pyperclip as clip
current_urls = urls[0:1]
for url in current_urls:
    webbrowser.get('open -a /Applications/Google\ Chrome.app %s').open(str(url))


webbrowser.get('open -a /Applications/Google\ Chrome.app %s').open(url)

### Analyzing HTML

categories = pd.read_csv("categories.csv")
categories_lookup = categories.set_index("champion").to_dict()['class']
categories_lookup['Ziggs']


def modify_champion(champion):
    if champion == "Kaisa":
        champion = "Kai'Sa"
    if champion == "KhaZix":
        champion = "Kha'Zix"
    return champion

team_id = 0
teams = []
roles = []

for filename in glob("html/*"):
    with open(filename) as f:
        text = f.read()
    first_regex = re.findall(">Picks<(.*?)</div>", text.replace("\n", " "))
    team1 = first_regex[0]
    team2 = first_regex[1]
    team1 = [x.split(" stats'>")[0] for x in team1.split("title=\'")[1:]]
    team2 = [x.split(" stats'>")[0] for x in team2.split("title=\'")[1:]]

    team_id += 1
    for champion in team1:
        champion = modify_champion(champion)
        teams.append(team_id)
        role = categories_lookup[champion.replace(" ", "_")]
        roles.append(role)
    team_id += 1
    for champion in team2:

        champion = modify_champion(champion)
        teams.append(team_id)
        role = categories_lookup[champion.replace(" ", "_")]
        roles.append(role)    
df = pd.DataFrame()
df['team'] = teams
df['role'] = roles
df


df.to_csv("final_output.csv", index=False)
