==> 5. Machine Learning New Features.ipynb <==
import pandas as pd
from pandasql import sqldf
import random


extract = pd.read_csv("final_output.csv")
extract


df = sqldf("""
SELECT
role1,
role2,
role3,
role4,
role5,
role6,
role7,
role8,
role9,
role10,
role11
FROM
    (SELECT
      team,
      SUM(CASE WHEN role = 1 THEN 1 ELSE 0 END) AS role1,
      SUM(CASE WHEN role = 2 THEN 1 ELSE 0 END) AS role2,
      SUM(CASE WHEN role = 3 THEN 1 ELSE 0 END) AS role3,
      SUM(CASE WHEN role = 4 THEN 1 ELSE 0 END) AS role4,
      SUM(CASE WHEN role = 5 THEN 1 ELSE 0 END) AS role5,
      SUM(CASE WHEN role = 6 THEN 1 ELSE 0 END) AS role6,
      SUM(CASE WHEN role = 7 THEN 1 ELSE 0 END) AS role7,
      SUM(CASE WHEN role = 8 THEN 1 ELSE 0 END) AS role8,
      SUM(CASE WHEN role = 9 THEN 1 ELSE 0 END) AS role9,
      SUM(CASE WHEN role = 10 THEN 1 ELSE 0 END) AS role10,
      SUM(CASE WHEN role = 11 THEN 1 ELSE 0 END) AS role11
    FROM extract
    GROUP BY 1
    ) a
""", locals())
df['target'] = "_0"
df



# I realize stratified sampling could work but doing this
for index, row in df.iterrows():
    roles = []
    for i in range(1, 12):
        if row[f'role{i}'] == 1:
            roles.append(i)
    random_choice = random.choice(roles)
    df.at[index, 'target'] = "_" + chr(random_choice+97)
    df.at[index, f'role{random_choice}'] = 0
    

def add_three_numbers(a, b, c):
    return a + b + c

df['tank_or_catcher'] = df[['role1', 'role7', 'role8']].apply(lambda x: add_three_numbers(*x), axis=1).astype(float)
# df.drop(['role1','role7','role8'], axis='columns', inplace=True)
df['diver_juggernaut_or_melee'] = df[['role2', 'role3', 'role9']].apply(lambda x: add_three_numbers(*x), axis=1).astype(float)
df.drop(['role2','role3','role9'], axis='columns', inplace=True)


df

from numpy.random import seed
from sklearn.model_selection import train_test_split
seed(32)
split_size = 0.4 if df.shape[0] < 100000 else 40000/df.shape[0]
train, temp = train_test_split(df, test_size=split_size)
dev, test = train_test_split(temp, test_size=0.50)
X_train = train.drop(['target'], axis=1)
y_train = train[['target']]
X_dev = dev.drop(['target'], axis=1)
y_dev = dev[['target']]
X_test = test.drop(['target'], axis=1)
y_test = test[['target']]
X = df.drop(['target'], axis=1)
y = df[['target']]
train.shape, dev.shape, test.shape, train.target.value_counts(normalize=True)


from sklearn.metrics import accuracy_score

def get_score():
    seed(32)
    model.fit(X_train, y_train)
    accuracy_train = accuracy_score(y_train, model.predict(X_train))
    accuracy_dev = accuracy_score(y_dev, model.predict(X_dev))
    print("Train Score:", accuracy_train)
    print("Dev Score:", accuracy_dev)
accuracy_baseline1 = accuracy_score(y_dev, [5]*X_dev.shape[0])
accuracy_baseline2 = accuracy_score(y_dev, [6]*X_dev.shape[0])
accuracy_baseline3 = accuracy_score(y_dev, [8]*X_dev.shape[0])
print("Baseline Score1:", accuracy_baseline1)
print("Baseline Score2:", accuracy_baseline2)
print("Baseline Score3:", accuracy_baseline3)

dev

from sklearn import linear_model, neighbors, tree, svm, ensemble
from xgboost import XGBClassifier
# Algorithms
algorithms = []
# algorithms.append(linear_model.LogisticRegression())
# algorithms.append(linear_model.SGDClassifier())
# algorithms.append(neighbors.KNeighborsClassifier(n_neighbors=1))
# algorithms.append(neighbors.KNeighborsClassifier(n_neighbors=3))
# algorithms.append(neighbors.KNeighborsClassifier(n_neighbors=5))
# algorithms.append(neighbors.KNeighborsClassifier(n_neighbors=7))
# algorithms.append(neighbors.KNeighborsClassifier(n_neighbors=11))
algorithms.append(tree.DecisionTreeClassifier(max_depth=3))
# algorithms.append(svm.SVC(kernel='linear'))
# algorithms.append(svm.SVC(kernel='poly'))
# algorithms.append(svm.SVC(kernel='rbf'))
# algorithms.append(svm.SVC(kernel='sigmoid'))
# algorithms.append(XGBClassifier(learning_rate=0.1, max_depth=10, min_child_weight=13, n_estimators=40, nthread=1, subsample=0.55))


for algorithm in algorithms:
    algorithm.fit(X_train, y_train)
    pred = algorithm.predict(X_dev)
    print(accuracy_score(pred, y_dev))
    

model = tree.DecisionTreeClassifier(max_depth=3)
model.fit(X_train, y_train)
pred = algorithm.predict(X_dev)
print(accuracy_score(pred, y_dev))

model = tree.DecisionTreeClassifier(max_depth=3)

model.fit(X, y)

import graphviz
import os

dot_data = tree.export_graphviz(model, out_file=None,
                      feature_names=train.drop(["target"], axis=1).columns,
                      class_names=['vanguard', 'diver', 'juggernaut', 'enchanter', 'adc', 'mage', 'warden', 'catcher', 'melee', 'battlemage', 'assassin'],
                      filled=True, rounded=True,
                      special_characters=True)
graph = graphviz.Source(dot_data)
graph.render()

os.remove("Source.gv")
os.rename("Source.gv.pdf", "decision_tree.pdf")


sorted(y.values), len(set(y))

y.value_counts()
